06_pjt_bank API/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
Django==4.2.11
django-environ==0.12.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
executing==2.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.59.7
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.3.0
wcwidth==0.2.13
Wikipedia-API==0.7.3
zipp==3.21.0
